refactor: move namespace management functions to namespace_manager.rs
- Moved NamespaceContext enum and impl from bundler.rs to namespace_manager.rs
- Moved require_namespace() and generate_required_namespaces() to namespace_manager.rs as free functions
- Reduced bundler.rs by ~190 lines (from 7276 to 7086 lines)
- Increased namespace_manager.rs by ~194 lines (now 1352 lines)
- Better separation of concerns: namespace-specific logic is now properly housed in namespace_manager.rs
- All tests passing, no clippy warnings

This addresses the concern about adding complexity to the already huge bundler.rs file.
